removes asyncstream handler for state updates (#46)
It may be that actor is VERY MUCH the wrong mechanism for encapsulating
a connection, but the "listener not becoming ready" may well be related
to not having an explicit handler in place to capture and track when it
does...

This _does_ assume that I can schedule a Receive() callback before
things are ready - I hope that's the case. We'll see.