fix build with Xcode 16 by disabling `InternalImportsByDefault`
`InternalImportsByDefault` makes all imports `internal` by default.
This means that none of those imports can be used in public interfaces.
Currently this repo _is_ using lots of `internal imports` and it seems
like the feature is ignored in Xcode 15. On Xcode 16 this breaks expectingly.

I'm not 100% clear on how the transition to this feature will work, and
if the feature can only be enabled once everyone's building with Xcode 16?

Disabling the feature now and adding a comment to get the Xcode 16 build
to work.