allow setting the logVerbosity of Documents created within the repo
`LogVerbosity` on automerge's Documents are internal and can only
be specified on document creation.
This PR allows setting the default for this via the repository.

Note: first approach was to add a case in `LogComponent`, though
that _also_ needs a logger. Given we don't really want to log into the
`Document`'s logger from the `Repo`, it felt better to add a separate
public variable.